+++ app/frontend/packs/javascript/components/each_step_graph_component.js
