+++ app/frontend/packs/stylesheets/_steps.scss
@@ -15,7 +15,7 @@
       font-size: 18px;
     }
   }
-  .each_step {
+  .step_tasks_box {
     max-width: 350px;
     padding-bottom: 50px;
     margin: 0 auto;
