+++ app/views/issues/show.html.slim
@@ -8,18 +8,18 @@
         .step_title_box
           h4.step_title
             | 1.DBを作成
-        .each_step
-          step-circle
+        .step_tasks_box
+          each-step
       .col-lg-4.col-md-12
         .step_title_box
           h4.step_title
             | 2.RubyでDBをラップして該当するクエリを投げる
-        .each_step
-          step-circle
+        .step_tasks_box
+          each-step
       .col-lg-4.col-md-12
         .step_title_box
           h4.step_title
             | 3.コマンドラインで操作できるインターフェイスを作成
-        .each_step
-          step-circle
+        .step_tasks_box
+          each-step
 = javascript_pack_tag 'javascript/steps_vue.js'
